% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_conversion.R
\name{ndc_convert}
\alias{ndc_convert}
\title{Convert NDCs from 10-digits to 11-digits}
\usage{
ndc_convert(path, var_name = "ndcnum")
}
\arguments{
\item{path}{A path (character) to the original csv file downloaded.}

\item{var_name}{A variable name for the converted NDCs in the output data
(Default: "ndcnum").}
}
\value{
Returns nothing. As a side effect, it creates a .dta file in the same folder.
}
\description{
Convert NDCs from 10-digits to 11-digits
}
\examples{
\dontrun{
ndc_convert("path/to/the/file")
}
}
